Make fixes according to review comments
