+++ testdata.py
@@ -1,0 +1,9 @@
+class TestData:
+    books = [
+        {"book": "Град обречённый", "genre": "Фантастика"},
+        {"book": "Что делать, если ваш кот хочет вас убить", "genre": "Ужасы"},
+        {"book": "Кто убил мистера Н", "genre": "Детективы"},
+        {"book": "Котёнок по имени Гав", "genre": "Мультфильмы"},
+        {"book": "Гордость и предубеждение и зомби", "genre": "Комедии"},
+    ]
+
